Refactor: DTO 패키지 구조 개선 -> request/response로 세분화
- DTO 패키지 구조 세분화
- 기존 : DTO/도메인명/DTO들
- 개선 : DTO/도메인명/request OR response/DTO들

request/response DTO가 아닌 것들은 도메인명 하위에 바로 작성
