+++ src/main/java/com/yangsunkue/suncar/dto/auth/request/SignUpRequestDto.java
@@ -1,4 +1,4 @@
-package com.yangsunkue.suncar.dto.auth;
+package com.yangsunkue.suncar.dto.auth.request;
 
 import com.yangsunkue.suncar.entity.user.User;
 import com.yangsunkue.suncar.common.enums.UserRole;
