+++ src/main/java/com/yangsunkue/suncar/dto/auth/response/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.yangsunkue.suncar.dto.auth;
+package com.yangsunkue.suncar.dto.auth.response;
 
 import com.yangsunkue.suncar.entity.user.User;
 import lombok.*;
