+++ src/main/java/com/yangsunkue/suncar/dto/car/response/RegisterCarResponseDto.java
@@ -1,4 +1,4 @@
-package com.yangsunkue.suncar.dto.car;
+package com.yangsunkue.suncar.dto.car.response;
 
 
 import lombok.Builder;
