+++ src/main/java/com/yangsunkue/suncar/dto/user/response/UserProfileResponseDto.java
@@ -1,4 +1,4 @@
-package com.yangsunkue.suncar.dto.user;
+package com.yangsunkue.suncar.dto.user.response;
 
 import com.yangsunkue.suncar.common.enums.UserRole;
 import com.yangsunkue.suncar.entity.user.User;
