+++ src/main/java/com/yangsunkue/suncar/service/facade/CarManagementServiceImpl.java
@@ -2,7 +2,7 @@ package com.yangsunkue.suncar.service.facade;
 
 import com.yangsunkue.suncar.common.enums.BrandName;
 import com.yangsunkue.suncar.dto.car.CarListResponseDto;
-import com.yangsunkue.suncar.dto.car.RegisterCarResponseDto;
+import com.yangsunkue.suncar.dto.car.response.RegisterCarResponseDto;
 import com.yangsunkue.suncar.repository.car.CarListingRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
